requirements.txt
numpy
dali-dataset
torch
torchaudio